+++ api/kogitoinfra_types.go
@@ -24,10 +24,10 @@ import (
 type KogitoInfraConditionType string
 
 const (
-	// SuccessInfraConditionType ...
-	SuccessInfraConditionType KogitoInfraConditionType = "Success"
-	// FailureInfraConditionType ...
-	FailureInfraConditionType KogitoInfraConditionType = "Failure"
+	// KogitoInfraSuccess ...
+	KogitoInfraSuccess KogitoInfraConditionType = "Success"
+	// KogitoInfraFailure ...
+	KogitoInfraFailure KogitoInfraConditionType = "Failure"
 )
 
 // KogitoInfraConditionReason describes the reasons for reconciliation failure
@@ -48,6 +48,8 @@ const (
 	ResourceConfigError KogitoInfraConditionReason = "ResourceConfigError"
 	// ResourceMissingResourceConfig related resource is missing a config information to continue
 	ResourceMissingResourceConfig KogitoInfraConditionReason = "ResourceMissingConfig"
+	// ResourceSuccessfullyConfigured ..
+	ResourceSuccessfullyConfigured KogitoInfraConditionReason = "ResourceSuccessfullyConfigured"
 )
 
 // KogitoInfraInterface ...
@@ -80,8 +82,8 @@ type ResourceInterface interface {
 
 // KogitoInfraStatusInterface ...
 type KogitoInfraStatusInterface interface {
-	GetCondition() KogitoInfraConditionInterface
-	SetCondition(condition KogitoInfraConditionInterface)
+	GetConditions() *[]metav1.Condition
+	SetConditions(conditions *[]metav1.Condition)
 	GetRuntimeProperties() RuntimePropertiesMap
 	AddRuntimeProperties(runtimeType RuntimeType, runtimeProperties RuntimePropertiesInterface)
 	GetVolumes() []KogitoInfraVolumeInterface
@@ -91,20 +93,6 @@ type KogitoInfraStatusInterface interface {
 // RuntimePropertiesMap defines the map that KogitoInfraStatus
 // will use to link the runtime to their variables.
 type RuntimePropertiesMap map[RuntimeType]RuntimePropertiesInterface
-
-// KogitoInfraConditionInterface ...
-type KogitoInfraConditionInterface interface {
-	GetType() KogitoInfraConditionType
-	SetType(infraConditionType KogitoInfraConditionType)
-	GetStatus() v1.ConditionStatus
-	SetStatus(status v1.ConditionStatus)
-	GetLastTransitionTime() metav1.Time
-	SetLastTransitionTime(lastTransitionTime metav1.Time)
-	GetMessage() string
-	SetMessage(message string)
-	GetReason() KogitoInfraConditionReason
-	SetReason(reason KogitoInfraConditionReason)
-}
 
 // ConfigVolumeSourceInterface ...
 type ConfigVolumeSourceInterface interface {