+++ RELEASE_NOTES.md
@@ -10,5 +10,6 @@
 - [KOGITO-4477](https://issues.redhat.com/browse/KOGITO-4477) - Operator sees the Infinispan as not ready after it is restarted
 - [KOGITO-4644](https://issues.redhat.com/browse/KOGITO-4644) - Operator Binary build & BDD tests should handle correctly fast jars
 - [KOGITO-4453](https://issues.redhat.com/browse/KOGITO-4453) - Operator overwrites user's configmap provided with propertiesConfigMap
+- [KOGITO-4469](https://issues.redhat.com/browse/KOGITO-4469) - Conditions should reflect latest evaluated status
 
 ## Known Issues